Run template script to create repo skeleton
