+++ .travis.yml
@@ -17,4 +17,4 @@ after_success:
 env:
   global:
   - SAUCE_USERNAME: Desire2Learn
-<%= deployInfo %>
+
