+++ locales/en.js
@@ -1,0 +1,3 @@
+export const val = {
+	'myLangTerm': 'I am a dynamically imported localized string!'
+};
